refactor(cart): 단위 금액 계산을 월단위에서 일단위로 변경하고, 근로 날짜에 해당하는 임금을 계산하는 방식으로 변환

+++ cartpost-service/src/main/java/com/example/cartpostservice/cart/controller/dto/response/CartItemsGetResponse.java
@@ -22,7 +22,7 @@ public record CartItemsGetResponse(
         String paymentType,
 
         @Schema(description = "결제 금액", example = "100000")
-        int amount
+        Long amount
 
 ) {}
